requirements.txt
tensorflow
pillow
streamlit
fastapi
uvicorn
